requirements-ci.txt
# Minimal CI dependencies - CPU only, no GPU-heavy packages
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.6.0
black>=24.0.0
codecov>=2.1.0
coverage>=7.0.0

# Core dependencies for basic OCR testing
pillow>=11.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.17.0
pdfminer.six>=20221105
requests>=2.32.0
numpy>=2.0.0
opencv-python-headless>=4.10.0

# Exclude heavy dependencies:
# - paddlepaddle (96MB)
# - torch/torchvision (GPU dependencies)
# - mistralai dependencies
# - langchain ecosystem